Bugfix Commit: node export d6 migrate had some bugs
The module did not work off the bat due to trying to directly access a
string as if it were an array. This change also includes a quick and
dirty hack that forces the body filter format to be 'filtered_html'.
This prevents an error and provides an easy temporary solution.