Add Module: libraries
This module is a dependency module for supporting and integrating 3rd
party modules without placing them inside of modules. This reduces
downloading the same copy of a 3rd party module multiple times and
avoids having to update and manage multiple copies of the same module.
A large number of projects depend on this module.