Bugfix: Validate existence of object property before calling
The following is an error on the calendar:

Notice: Undefined property: stdClass::$id in calendar_build_nodes()
(line 663 of /var/www/branches/kday/sites/all/modules/calendar/calendar.module).

That warning is a result of the module developers not validating the
variables that come from outside the function.
The actual problem might be the result of another module improperly
defining or using a particular field.
However, looking at the property itself 'id', I believe the calendar
module may be improperly assuming that the id object variable always
exists.

The immediate solution is to silence the warning is to ensure that the
property exists before calling it, but keep in mind that this might be
hiding the real problem that is caused elsewhere.