Revert "Add Module: cf 7.x-1.0"
This reverts commit 5d6f483b369f7a7101a66156274c1cd005b78654.

This should not be on master at this time.
I was testing out a 1 word alias to completely merge a local branch into
master without typing in 6 separate commands.
The process seems to have worked, but I ended up including files that
should not be pushed to master at this time.