Revert "Update Module: killfile 7.x-1.0"
This reverts commit d7a25d3888f269ab45f7c531861db68be7d43163.

This should not be on master at this time.
I was testing out a 1 word alias to completely merge a local branch into
master without typing in 6 separate commands.
The process seems to have worked, but I ended up including files that
should not be pushed to master at this time.