Revert "Add Module: 7.x-1.0-beta1"
This reverts commit 37759155b85832e018a6cc98f382f2bd642461b8.

This should not be on master at this time.
I was testing out a 1 word alias to completely merge a local branch into
master without typing in 6 separate commands.
The process seems to have worked, but I ended up including files that
should not be pushed to master at this time.