Bugfix: Fix invalid reporting as '' database
This is needed to ensure proper functionality with any database.
The problem is that the $db_query global from D6 is being used, but
it no longer exists in D7; instead, the drupal database function
'db_driver()' should be used (which was added in D7). This patch
changes the behavior to follow drupal API.

Upstream has been reported here: http://drupal.org/node/1105816