Bugfix: Fix invalid reporting as 'mysql' database
This is needed to ensure proper functionality on a non-mysql database.
The problem is that 'mysql' is being explicitly defined instead of using
the drupal database function 'db_driver()'. This patch changes the
behavior to follow drupal API.

Upstream has been reported here:
 - http://drupal.org/node/1105830
 - http://drupal.org/node/765364

This patch is only valid for -alpha1. for -beta3 and beyond, look into
the patch supplied at the second upstream link above.