Bugfix: Validate variables before operating on them
The first problem this solves is that of accessing an array item without
validating that the item exists. In php, this produces an error. isset()
us used here to prevent that from happending.

The second problem is that if the renderred variable is empty, then do
we display the block or not? The proper solution to this is to provide
options in the theme settings to toggle which variable should and should
not have its html wrapper displayed or not when the renderred variable
is empty. At this time we need to have the renderred variable not
displayed so html wrapper elements (such as divs) will not be produced.

A better long-term solution to the second problem is to provide the
mentioned theme settings variable empty toggle option.

This is only fixed in the page.tpl.php template for now, but the other
*.tpl.php files probably need the same or similar fix.