Workaround: Don't operate on non-existent objects
The following error appears when viewing an ical feed:
   Notice: Trying to get property of non-object in
   template_preprocess_date_vcalendar() (line 207 of
   sites/all/modules/date/date_views/theme/theme.inc).

The code seems to assume $vars['rows'] is an array.
This produces the above error when trying to operate on
$vars['rows'][0]->calendar_all_day when $vars['rows'] is NULL.

This fixes the symptom, but the real issue is why is $vars['rows'] null?
Should it be allowed to be NULL, if so then this is a proper fix.
Otherwise, this is workaround and more research is required to solve the
problem.

Given that I do not know either way, I am going to call this a
workaround for the time being.