Bugfix: Prevent css from making a link stop working
It seems that having a div that is a parent to a link of an image whose position is absolute will cause the link on the image to not work in all but Opera.
This changes the positioning to be on the image itself, which appears to solve the problem.
This also removes stale css.