Update: Add more printer-friendly specific css
Use the @page from CSS-3 to define default margins, be sure to test this in IE.
Force the default body background to be white.
Remove fields that should not be printed such as the calendar event group select list and the ical feed icon.
Make sure that the event date that is in the breadcrumb is still visible on the printer-friendly page even though the breadcrumb should not be visible.

issue: www-43