Workaround: some <div> blocks are vanishing
I noticed in my recent browser tests that the Skip to main content link started appearing in different browsers without being selected.
After testing, I discovered that there the parent <div> is completely vanishing out of the html on non-mozilla based browser.
It appears that there is something on the system specifically altering the HTML before it makes it to the browser or the browsers are altering the html themselves.
My workaround is to wrap the div inside of a span which then causes the span to vanish.