Bugfix: the submit link was out of line
That is, the submit link, such as the cancel button on the delete page was noticably offset from the reset of the submit buttons.
The problem was with the vertical-align being set to top instead of baseline.

Also, remove the underline so that it looks exactly like the other buttons (or as close as possible).