Bugfix: drupal cores base.css improperly uses css causing random breakage
The following css is impropery:
  .container-inline div {
    display: inline;
  }

What this does is cause every single div after the .container-inline to be altered to an inline element.
Instead, they should use something like
  .container-inline div.inline {
    display: inline;
  }

or

  div.container-inline {
    display: inline;
  }

This prevents the inline from being applied to divs that should not be applied to.

Because they do not supply a div.inline or div.container-inline, there is no complete solution.
Therefore, individual cases where the div should be inline needs to be explicitly handled.

This fixes problems with opera, safari, chrom, and seamonkey.