Update: Add the html id 'ckeditor-wysiwygarea' to the embeded html
This is required to get css styling specific to the ckeditor wysiwyg area to properly function.
As far as I can tell, there was no provided means to do this other than editing the source of ckeditor.
Therefore this step has to be re-performed when this module gets updated.

Here is how to do what I did:
1) modify ckeditor/_source/plugins/wysiwygarea/plugin.js (adding id="ckeditor-wysiwygarea")
2) re-package/compile the ckeditor _source into the ckeditor/ckeditor.js:
  a) Get the CKPackager from the ckeditor website:
    - http://docs.cksource.com/CKEditor_3.x/Developers_Guide/CKPackager
    - http://svn.ckeditor.com/CKPackager/trunk/bin/ckpackager.jar
  b) From inside the ckeditor directory, execute the packager:
    - java -jar ckpackager.jar ckeditor.pack

This is necessary to help solve issue www-26, but it also provides us with a means to make ckeditor more precicely match what the content as it is shown (while editing).

Issue: www-26

This is a re-apply of the commits cae87fbace9de8815915f5c50743bcab9b1b1635 and 6a94e2e72a6cbe9b9aa96051121c8f6c87a3e8b3 against ckeditor 3.6.