Workaround: Prevent invalid data from crashing php on save
This workaround is a bugfix for when invalid data is passed to certain functions.
Once again, instead of properly validating data, they make assumptions which lead to errors and in this case a complete crash.

While preventing damage is very important, I would also like to stop the cause of the damage.
As I do not know what is creating the invalid data, I am calling this a workaround.