Bugfix: Fix search box input button issues
This removes the "Search" text from the search box search button Using proper drupal hooks.
This is necessary to fix some bugs when the css is set to hide the search button text.
This uses the html title attribute (aka tooltip) to ensure that screen-readers and other text-based viewers can still identify what the button is.

This fixes the following problems:
- mozilla based browsers (specific versions&frontends, such as seamonkey and fedora 15s firefox) will randomly stop drawing pieces of a page, producing lots of random whitespace.
- webkit based browsers (specific frontends, such as midori) will always display the input button text, even if its hidden via the css!