Bugfix: Attempt to fix git bug
Somehow sites/all/themes/genesis_mcneese was added as a submodule when it should not have been.
Perhaps this is some feature of git add -A when some subdirectory has its own .git directory?
Either way, this is unwanted, so try and delete the genesis_mcneese directory so it can be re-added.