Bugfix: allow custom urls to be used, such as ssl:// and tls://
Simply changing the port to 443 is not enough to get a remote https page.
The ssl:// must be passed such that the socket is opened via an ssl packet.

When getting an http request, the $server argument was passed, but passing ssl://someurl produced an invalid http request.
This change ensures that only the url and not the ssl:// (or equivalent) is used in the http packet.
This is done in a generic way so that other protocols, such as udp://, are supported.