Bugfix: remove all the 'node' variable sql selects
As far as I can tell this appears to be a misuse of the sql and the query->add_field() function.

The problem is noticed when using queries with fields that have multiple values, with aggregation off.
Of those fields, some where taxonomy fields and I am intentionally not using a view relationship for the taxonomy because I do not want multiple (almost) duplicate rows.

While trying to figure out why the taxonomy field (called field_group in this case) was not returning any data when it should be, I noticed that the SQL query was screwed up.
All over the SELECT statements are selects such as (see SQL Query below):
  'node' AS field_data_field_url_path_node_entity_type

My first thought was why is there a select 'variable' being used, and then repeatedly for each field type that supports multiple values?
Looking at the JOINS, things get even worse.
All over the joins are things like (see SQL Query Results below):
  field_data_field_group.entity_type = 'node'

Looking at the actual database structure, it seems that field_data_field_group.entity_type = 'node' is a valid test.
If so, then that seems to make the SELECT variables for 'node' completely redundant at best.
Looking at the SQL Query Results there are 7 columns of useless 'node' data for every row returned.
If the entity_type is to be used, then a simple select on the entity types should be fine.

I isolated the problem to the this block of code:
  $entity_type_alias = $this->definition['table'] . '_' . $entity_type . '_entity_type';
  $this->aliases['entity_type'] = $this->query->add_field(NULL, "'$entity_type'", $entity_type_alias);

The paramaters are ($table, $field, $alias).
This means that there is no table field, the field being used is called 'node' (including the single quotes), and the alias is that long blob of text.
This is bizarre.
The database defines the entity_type so either the entity_type should be removed entirely, or the actual database columns called entity_type should be used.
Removing them reveals that those fields are used and so that leaves with the only choice of using the correct data.

The above block of code is then replaced with the following block of code:
  $this->aliases['entity_type'] = $this->query->add_field($this->definition['table'], 'entity_type');

This appears to make sense to me, but once this is done a regression happens:
  PHP Fatal error:  Class name must be a valid object or a string in /var/www/development/includes/common.inc on line 7618, referer: https://wwwdev.mcneese.edu/admin/structure/views/view/more_pages/edit/more_pages_news_events_and_more

Thats a fatal error, so the website shows a WSOD.
Further investigation shows that somehow an empty field is being defined an attempted to being used.
This empty field is, well, empty.
What this means is that there is no valid object or string because it does not exist!

I cannot figure out where or how this empty field is being inserted, so the only thing I can do is ensure that the empty field never gets processed.

This patch incorporates the bugfix and the empty field sanity checks.

SQL Query:
  SELECT DISTINCT node.nid AS nid, node.title AS node_title, field_data_field_event_datetime.delta AS field_data_field_event_datetime_delta, field_data_field_event_datetime.language AS field_data_field_event_datetime_language, field_data_field_event_datetime.bundle AS field_data_field_event_datetime_bundle, field_data_field_event_datetime.field_event_datetime_value AS field_data_field_event_datetime_field_event_datetime_value, field_data_field_event_datetime.field_event_datetime_value2 AS field_data_field_event_datetime_field_event_datetime_value2, field_data_field_event_datetime.field_event_datetime_rrule AS field_data_field_event_datetime_field_event_datetime_rrule, field_data_field_order_date.field_order_date_value AS field_data_field_order_date_field_order_date_value, 'node' AS field_data_field_url_path_node_entity_type, 'node' AS field_data_field_event_datetime_node_entity_type, 'node' AS field_data_field_featured_image_node_entity_type, 'node' AS field_data_field_event_image_node_entity_type, 'node' AS field_data_field_spotlight_image_node_entity_type, 'node' AS field_data_field_headline_node_entity_type, 'node' AS field_data_body_node_entity_type
  FROM
  node node