Update: make the menu item menu list unambiguous
Users may specify multiple menus of the same name.
Displaying only the name in a select list makes it unlikely for the user to know which is the correct menu.
This adds two new tags: url and id to the select list.

Also, it may be good idea to later on provide an interface where the site admin can toggle this behavior.