Bugfix: Fix node load
The commit: fe2aa03a1688061efce3777a63f10148c631fc9c broken the node_load.
The problem was not noticed because the cache was not cleared after making the change.

It seems the 'load arguments' must access (vid, vid) instead of (nid, vid).
This seems wrong and I suspect I might still be mis-using this.
For now, revert this part of the said commit to ensure functionality.