Bugfix: convert %20 to spaces in active_path
Workbench menu supports spaces in the url path.
However, the url is stored in the proper way when stored in the database.
That is to say that it uses things like %20 instead of a space.
The problem is that when the active path from the database is compared
with the current active path in the browser, the '%20' is not equal to ' ' during the comparison tests.

This converts the '%20' in the active path to spaces.

There are numbers of other characters other than spaces with this issue, but those are not supported yet.