Update Module: views 7.x-3.x-dev (2012-Jun-09)
The development release fixes a lot of major problems that was plaguing our websites.
Many of the fixes or workarounds I made were either accepted and applied upstream or properly fixed.
Not all of the matches have been updated and the following commits need to be reviewed to see if they are necessary and update them to apply with the no be reviewed to see if they are necessary and update them to apply with the newer code:
- d6fb944 Bugfix: remove all the 'node' variable sql selects
- 5d4e3de Feature: Support case-insensitive searches in views module