Bugfix: have nav fallback to div
while it makes sense to use nav in place of a list, it seems that nav is not considered something that can directly hold lists according to the standard.
The clients must always wrap the content of the nav in the list instead of having nav directly replace the list.