Bugfix: remove/replace problematic variables and add new ones
There are variables assigned to each of the menus.
The way in which these are calculated was not ideal and lead to problems.

This removes the most problematic ones.

This adds an additional variable to workbench_menu_list() so that custom implementations will know if they are dealing with a topmost or a child list.