Bugfix: Do not allow menu items to go missing
Some menu items were not appearing due to another bug that needs to be investigated.
The situation is that some menu items end up in the database with duplicate settings.
Only one of these items gets displayed because the id already exists.

It seems that duplicate menu_id's with different parent ids cause this problem.