Bugfix: prevent possible install time problems with cf-2.x
This module was originally written for cf-1.x.
It has been modified to be compatible with cf-2.x's pseudo-backwards compatibility.
It turns out that during install, if the cf-2.x backwards compatibility is not enabled, it is then possible to create a WSOD.
To prevent this, a lot of the old cf-1.x style functions are being wrapped in conditions to see if cf-2.x is in use.