Bugfix: handle error case when unable to connect to database
A good amount of the theming code depends on things that only work when the database is available.
When the database is unavailable, certain functions fail to become defined and a big nasty broken error page is presented.

To avoid this, a static version of maintenance.tpl.php is provided.
Certain functions also have to be wrapped with existence tests to avoid white screens.