Bugfix: fix problems with handling of image styles
The stream wrappers defined here will pretend to be valid for a given image style.
The actual work is then handed off to the public:// or private:// stream wrappers.

Different situations were not being handled, such as deleting of generated image styles.
This also does a check to see if the generated style already exists before generating it (improving performance).

This also fixes problems with url_stat().
url_stat() for the stream wrappers has to return the proper file mode so that php functions, like is_file(), are able to properly handle the uri.