Bugfix: rewrite mcneese_file_db, fixing logic flaws and improving performance
There were a few problems in the logic when reading and writing files.
The buffer size has been increased to 131072.
For this to be effective, the function (add in php 5.4) called stream_set_chunk_size() is used.
Without this function, all reads and writes are forced t 8192 bytes (it does not matter what the stream_set_read_buffer() or stream_set_write_buffer() are set to).