Update: improve user-friendliness with touch interfaces
Touch interfaces mouse-click (really finger-tap) can cause the button on the just opened menu to be clicked on.
This breaks the functionality.
The float regions now all have titles and adjust ~40px to the left or right when locked down.
This makes it easier to finger-tap open menus and then finger-tap close them while still supporting mouse-over effects.
The mouse overs do not apply the ~40px adjustment.