Bugfix: do not assign default when multiple groups are available
The module term_ref_autocomplete has a coded-in default for the workbench access module fields.
This is a problem because it is desired that the default to not be set.

However, for user-friendliness, a default could and will be set for users who are only assigned to 1 group.

This changes the behavior to only assign a default value when the user is assigned to only 1 group.