Bugfix: items don't always show attached children
There was a logic flaw in how the menu items were being populated.
When a child is processed, it will add itself to the parents 'children' array.
If the parent has not yet been processed and attached, then it overrides the pre-populated 'children' array.
The solution is to check and see if the array is already defined and if so, do not overwrite it.