Bugfix: fix logic flaw introduced in previous active path detection
The previous change in the active path detection code utilized the drupal current_path() function to see if the current path is a sub-path for some node.
This worked, but the design allowed for including and marking other menu items as active when they should not be.

This changes the logic to check to see if the current path for the node matches the current url alias path.