Bugfix: fix problems with workbench_menu_access_menu() queries
There was a logic flaw in the workbench_menu_access_menu() function.
The standard join() seems to be doing an inner_join() and I was thinking it was a left_join().

But just to be on the safe side, I decided to use two separate queries.
The first is for the query against the groups the user is allowed to access.
The second is for the query the groups the roles in which the user is in is allowed to access.