Hack: implement print font-size and line-height exceptions
This is very much not the best way to do this, but it is a compatible and simple way to do this.

This provides font-sizes and line-height dimensions for pixels of sizes 1-99 for print only media types.
The idea is that it may be necessary to assign a custom font-size and line-height for specific cases only when printing.

Doing it this way allows me to make this functionality available to users while still denying them access to write css themselves (for security reasons).

www-698