Bugfix: fix webform label length issues (warning, read more..)
The label is an accessibility requirement, but the current setup puts too many restrictions on a label to make it usable in all cases.

This patch changes the behavior to enable longer labels so that markup and long sentences may be used.

Warning:
This patch changes the database structure unofficially while there is an upstream bug report for this.
This will cause a problem on update, in the event that they create their own update function (for other reasons) or to fix this problem.
Updating the module will require delicate handling to ensure no updates are missed.

see: https://drupal.org/node/342722

www-708