Revert "Update: mark required form input as required"
This reverts commit 2dd2f7950db853dd30288b0c513b722b83d93cf1.

This does not work well automatically at this time.
A better implementation is needed.

This caused a bug where forms did not submit due to the presence of conditionally required fields.
This also caused problems with hidden required fields.
That is fields that are hidden inside of a collapsed fieldset.