Bugfix: active trail was not being set when the current path was not explicitly defined in the menu
When a particular path is a sub-path of one or more menus, then the menu is presented.
This changes the breadcrumb, but if the current path has no menu item within the menu, then the breadcrumb was not being properly set.

The behavior should be to, at the very least, assign the breadcrumb if the trail is valid.