Bugfix: prevent IE from having a heart attack because it doesn't support a property
Internet Explorers javascript processing code will stop functioning entirely when it doesn't support a property.
Normal behavior would be just to exit out the function, exit out of the script, or skip over the unsupported feature.
In the case with IE, it crashes every javascript script.

To prevent this from happening, wrap the call in an if condition to check to see if the property first exists before calling.