Update: major improvements, performance, initial restricted support, different database structure
This improves performance by only accessing the database when needed during setUri() and similar function calls.

This adds the initial support for restricted file database.
The basics are in place, but much testing is needed.
Additionaly work is probably necessary.

The change forces a database structural change such that:
- mcneese_file_db_associations is split into two different databases:
  - mcneese_file_db_unrestricted
  - mcneese_file_db_restricted

The database must not be migrated from the old to the new.