Update: Rework create content handling pages (warning, read notes)
This requires a patch to the drupal core node add pages for creating content.
The McNeese change was to get rid of the caching in and looping in the node_menu() function.
The desired commit is named: "rework node menu generation" for the core node module.

If the core patch is not applied, be sure to revert this commit to prevent error messages and other problems.

Note: I thought theere was a jira ticket for this, but I cannot find it at this time.