Update: major updates, bugfixes, and features for the mcneese templates
This adds a new database column for storing hidden values associated with a template that do not appear on the node itself.
This adds features to create template input target arrays.
This improves the node save and node load process.
Use trim() on taxonomy id before checking to see if it is an integer,