Update: redesign template management to allow for multiple fields and other changes
This primarily redesigns the template system to allow for cardinality.
This adds support for more field types and changes the overall field structure.
The templates are re-designed and requires that all templates be re-created.

The interface has been cleaned up slighty (more work needs to be done).
The documentation has been updated accordingly.

Some field types are not yet support and others, like media, do not yet have support for special handling, such as the alt and title text for image media.