Feature: add support for workbench menu aliases
This will hopefully replace the workbench menu item "menu".
The older type of menu was too complicated and buggy.
This simplifies the logic by using aliases in a way that is compatible with the current database structure.

In the future, the paths should just point to menu ids in the database.
This would allow for any number of aliases and allow for them to point to the same table without any extra fields.

The workbench menu item menu item may also be removed in the future.