Bugfix: improve file handling, fix temporary file handling, add caching
This requires  a change in drupal core to allow for streamwrappers to function without directories.
- Thus, the static protected $directoryless = TRUE; is defined.

There is a bug where when temporary files are uploaded but do not become permanent because they already exist, drupal might accidentally remove the original, permanent, file when deleting the temporary file.

The database should be checked to see if the file is already uploaded when making changes.

Some places, namely the upload process, will repeatedly perform fresh file operations.
These operations require loading the file information, which requires a database call.
Avoid this database call by caching the file meta data per php execution.
During the upload processed, this saves ~31 database reads.
This could have other improvements elsewhere, but only the upload process was tested.

Add a simulated/approximate file mode for the stat system to use.
This also allows for properly reporting that a given file is a file (and not some directory or special file type).

Cleanup some of the code to not define a variable = FALSE and then return that variable when FALSE can be directly returned (generally for error handling).
This should have a trivial performance benefit by avoiding unnecessary allocations.