Bugfix: under certain circumstances empty breadcrumbs appear
Fixing this required using menu_get_item() to get the router_path.
Some of the existing code uses this, so there is a possibility that code can be further simplified.
Rather than take the risk of trying to clean up and optimizing the code, I am focusing on stability and making no further changes.