Feature: wrap install and uninstall core functionality in transactions
This prevents bugs or problems during the install or uninstall process from partially adding or partially removing database information.

Before this patch, if a module were to have a problem install 1000 tables about half way through, there would be no way to uninstall the module outside of manually dropping ~500 tables.
Now transactions are used during the install and uninstall process so that if something went wrong, no changes are made.