Update: rework and fix upload api to handle corner cases and allow for multiple file uploads
The file upload for special cases, such as webform, did their own custom file save that caused the upload to not be properly renamed.
This significantly reworks the design to compensate for those cases without having to change the problematic modules.

The multiple file uploads of the same type was something I wanted to build and solving this happened to lead to adding this functionality.
Multuple file uploads are multiple uploads of the same file, but only one instance exists in the database.