Feature: add support for facebook js for the bookstore website
To prevent (some) potential security issues, provide the javascript locally.
This allows the site to restrict the '<script>' tag from users, among other things.

For the bookstore, a block is require to be created in the 'hidden' block of the mcneese_www theme.
The code looks like:
<?php
  if (function_exists('mcneese_third_party_add_facebook_connect_legacy_js')) {
    mcneese_third_party_add_facebook_connect_legacy_js();

    if (function_exists('mcneese_third_party_do_facebook_connect_js')) {
      mcneese_third_party_do_facebook_connect_js('71065ce9bc688916c2ab0496f7a6db0f');
    }
  }
?>

Currently, the legacy is used.