Security: restrict cron operations to unauthorized, non-localhost access
It seems that there is some pre-processing done in hook_init().
hook_init() is run on every single page.
There is no reason to do this for users not authorized to run cron jobs.

Add additional permission checks to the ultimate_cron_run_launchers() function to ensure that cron jobs can only be called by authorized users who are not localhost.

Make sure localhost is still able to perform tasks because it runs as anonymous.